requirements.txt
groq==0.3.1
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
rich>=13.7.0
google-api-python-client>=2.0.0 